refactored filtering
